Refactor cover page and add operations page with election details display
