+++ PVB-Frontend/src/Pages/BackOffice.js
@@ -47,7 +47,6 @@ const PVB_CoverUI = () => {
                     </button>
                 </div>
             </div>
-
         </div>
     );
 }
