--- PVB-Frontend/src/Resources/coverPage.css
@@ -1,54 +1,0 @@
-.coverPage::before {
-  transform: scaleX(0);
-  transform-origin: bottom right;
-}
-
-.coverPage:hover::before {
-  transform: scaleX(1);
-  transform-origin: bottom left;
-}
-
-.container {
-  text-align: center;
-  margin-top: 50px;
-}
-
-.centered {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.coverPage::before {
-  content: " ";
-  display: block;
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  inset: 0 0 0 0;
-  background: hsl(200, 88%, 84%);
-  z-index: -1;
-  transition: transform .3s ease;
-}
-
-.coverPage {
-  position: relative;
-  font-size: 5rem;
-}
-
-html {
-  block-size: 100%;
-  inline-size: 100%;
-}
-
-body {
-  min-block-size: 100%;
-  min-inline-size: 100%;
-  margin: 0;
-  box-sizing: border-box;
-  display: grid;
-  place-content: center;
-  font-family: system-ui, sans-serif;
-}
